+++ generators/app/gen/backend-node-express.js
@@ -14,6 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+const fs = require('fs');
+const fsExtra = require('fs-extra');
+const path = require('path');
 const sanitizeFilename = require('sanitize-filename');
 
 // information about that generator
@@ -21,6 +24,9 @@ exports.about = {
     displayName: 'Backend (Node - Express & React)',
     icon: '🛠',
 };
+
+const optionDatabaseMongo = 'Mongo';
+const optionDatabaseTypeORM = 'TypeORM';
 
 /**
  * A generator for Node.js based APIs (Express).
@@ -39,18 +45,141 @@ exports.run = async function () {
         return;
     }
 
-    const projectNameLower = projectName.toLowerCase();
+    const databases = await this.tools.promptMultiSelect(
+        'Select the databases, the backend supports:',
+        [{
+            name: optionDatabaseTypeORM,
+            checked: true
+        }, {
+            name: optionDatabaseMongo,
+            checked: false
+        }]
+    );
+
+    const options = {
+        mongo: databases.includes(optionDatabaseMongo),
+        name: sanitizeFilename(projectName.toLowerCase()),
+        typeORM: databases.includes(optionDatabaseTypeORM),
+    };
 
     // create output directory
     const outDir = this.tools
-        .mkDestinationDir(projectNameLower);
+        .mkDestinationDir(options.name);
+
+    const deleteFile = async (relPath) => {
+        const file = path.join(outDir, relPath);
+
+        await fs.promises.unlink(file);
+    };
+
+    const deleteFolder = async (relPath) => {
+        const folder = path.join(outDir, relPath);
+
+        await fsExtra.remove(folder);
+    };
+
+    const editFile = async (relPath, action) => {
+        const file = path.join(outDir, relPath);
+    
+        const newText = await action(
+            await fs.promises.readFile(file, 'utf8')
+        );
+    
+        await fs.promises.writeFile(file, newText, 'utf8');
+    };
+
+    const editJSON = async (relPath, action) => {
+        const file = path.join(outDir, relPath);
+
+        const obj = JSON.parse(
+            await fs.promises.readFile(file, 'utf8')
+        );
+    
+        await action(obj);
+    
+        await fs.promises.writeFile(file, JSON.stringify(obj, null, 4), 'utf8');
+    };
 
     const filesToOpenInVSCode = [];
 
+    // copy files
     await this.tools.withSpinner('Copying files', async (spinner) => {
         const filesToExclude = [];
 
         this.tools.copy(templateDir, outDir, null, filesToExclude);
+    });
+
+    // update package.json files
+    await this.tools.withSpinner('Update package.json files', async (spinner) => {
+        const packageJSONFiles = {
+            'package.json': async (packageJSON) => {
+                packageJSON.name = `${options.name}`;
+            },
+            'backend/package.json': async (packageJSON) => {
+                packageJSON.name = `${options.name}-backend`;
+
+                if (!options.typeORM) {
+                    delete packageJSON.dependencies['typeorm'];
+                    delete packageJSON.dependencies['pq'];
+                    delete packageJSON.devDependencies['@types/pg'];
+                    delete packageJSON.scripts['migration:create'];
+                }
+                if (!options.mongo) {
+                    delete packageJSON.dependencies['mongoose'];
+                    delete packageJSON.dependencies['@types/mongoose'];
+                }
+            },
+            'frontend/package.json': async (packageJSON) => {
+                packageJSON.name = `${options.name}-frontend`;
+            },
+        };
+
+        for (const file in packageJSONFiles) {
+            await editJSON(file, packageJSONFiles[file]);
+        }
+    });
+
+    await this.tools.withSpinner('Cleanups', async (spinner) => {
+        if (!options.typeORM) {
+            await deleteFile('backend/ormconfig.js');
+            await deleteFile('backend/src/database/typeorm.ts');
+
+            await deleteFolder('backend/src/database/entity');
+            await deleteFolder('backend/src/database/migration');
+        }
+
+        if (!options.mongo) {
+            await deleteFile('backend/src/database/mongo.ts');
+        }
+
+        const databaseIndexExports = [];
+        if (options.mongo) {
+            databaseIndexExports.push("export * from './mongo';");
+        }
+        if (options.typeORM) {
+            databaseIndexExports.push("export * from './typeorm';");
+        }
+
+        if (databaseIndexExports.length) {
+            await editFile('backend/src/database/index.ts', async () => {
+                return databaseIndexExports.join('\n');
+            });
+        } else {
+            await deleteFolder('backend/src/database');
+        }
+    });
+
+    // npm install
+    await this.tools.withSpinner(`Run 'npm install'`, async (spinner) => {
+        const nodeDirs = [
+            outDir,
+            path.join(outDir, 'backend'),
+            path.join(outDir, 'frontend')
+        ];
+
+        for (const dir of nodeDirs) {
+            this.tools.runNPMInstall(dir);
+        }
     });
 
     await this.tools