+++ generators/app/templates/backend-node-express/backend/src/database/typeorm.ts
@@ -1,5 +1,4 @@
 import * as path from 'path';
-import { connect as openMongoConnection, Mongoose } from 'mongoose';
 import { getConnectionManager as getTypeORMConnectionManager, Connection as TypeORMConnection } from 'typeorm';
 import { tick } from '../utils';
 import { IS_LOCAL_DEV, SCRIPT_EXT } from '../constants';
@@ -16,16 +15,6 @@ export const DEFAULT_TYPEORM_CONNECTION = 'default';
  */
 export type WithDBAction<TResult extends any = any> =
     (conn: TypeORMConnection) => TResult | PromiseLike<TResult>;
-
-/**
- * Describes an action for a 'withDB()' function call.
- *
- * @param {Mongoose} conn The current connection.
- *
- * @returns {TResult|PromiseLike<TResult>} The result of the action.
- */
-export type WithMongoAction<TResult extends any = any> =
-    (conn: Mongoose) => TResult | PromiseLike<TResult>;
 
 /**
  * Returns a new TypeORM connection to the application database.
@@ -97,31 +86,6 @@ export async function withDB<TResult extends any = any>(
     return tick(async () => {
         if (!connection) {
             connection = await getDBConnection();
-        }
-
-        return action(connection);
-    });
-}
-
-/**
- * Invokes an action for an open Mongo connection, and closes it after it is not used anymore.
- *
- * @param {WithMongoAction<TResult>} action The action to invoke.
- * @param {Mongoose} [connection] The custom connection to use.
- *
- * @returns {Promise<TResult>} The promise with the result of the action.
- */
-export async function withMongo<TResult extends any = any>(
-    action: WithMongoAction<TResult>, connection?: Nilable<Mongoose>
-): Promise<TResult> {
-    return tick(async () => {
-        if (!connection) {
-            connection = await openMongoConnection(process.env.MONGO_CONNECTION!, {
-                useNewUrlParser: true,
-                useUnifiedTopology: true,
-                useFindAndModify: false,
-                useCreateIndex: true
-            });
         }
 
         return action(connection);